backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
sqlmodel==0.0.25
python-dotenv==1.1.1
python-jose==3.5.0
slowapi==0.1.9
passlib==1.7.4
pydantic==2.12.0
bcrypt==5.0.0